Don't use keyword args in the fallback decode path, use positional args instead.
In the past codecs.decode and the decoder functions returned by
codecs.getdecoder only took positional arguments. codecs.decode now does take
keyword arguments but the underlying decoder functions may still not. The
default fallback decoder, latin_1_decode (representing iso-8859-1) is such
a decoder and so when the fallback decode path is triggered an exception
will hit. To fix this, simply pass the error mode string as a positional.